Allow chip to shutdown when driver is unloaded
When driver is loaded, GPIO reset is set as output and is controlled by driver.
When, driver is unloaded, GPIO become an input and by default, Raspberry applies
pull-up to it.

However, on unload, driver ask to wf200 chipset to enter very low consumption
state. WF200 will keep this state while reset is down. So, pull-up on this GPIO
breaks this behavior. Current patch will ask to GPIO driver to remove
pull-up/pull-down (and pull-up/pull-down configuration is not touched by
GPIO driver when GPIO is released). Since WF200 has a small pull-down built-in,
this configuration is sufficient to keep chip asleep.

In add, for SDIO mode, GPIO is controlled by mmc-pwrseq-simple driver. This
driver set reset GPIO on unload. This behaviour does not allows chip to stay in
low consumption mode. So, in SDIO, the correctly to entry low consuption mode is
to unload wfx driver, but keep mmc-pwrseq-simple enabled:

   rmmod wfx

This patch also apply no-pull-up to SDIO overlay, but it is only to keep it sync
with SPI overlay.