fix(wfx_test_agent): correct readout of driver version
The current upstream wfx driver does not use the `MODULE_VERSION`
variable anymore to set the version (as opposed to the legacy code
https://github.com/SiliconLabs/wfx-linux-driver/blob/15a127a0d112ce91729d755fa53b4e835a3a4acb/main.c#L46).
This can make the `wfx_test_agent read_driver_version` command fail.
Correct the command to fall back to the kernel-generated `vermagic`
field if `version` is not found.